separate SOTabBarSetting to a different file
