+++ SOTabBar/Classes/SOTabBarViewController.swift
@@ -82,23 +82,6 @@ open class SOTabBarViewController: UIViewController, SOTabBarDelegate {
     }
 }
 
-
-// Here you can customize the tab bar to meet your neededs
-
-public struct SOTabBarSetting {
-    
-    public static var tabBarHeight: CGFloat = 66
-    public static var tabBarTintColor: UIColor = UIColor(red: 250/255, green: 51/255, blue: 24/255, alpha: 1)
-    public static var tabBarBackground: UIColor = UIColor.white
-    public static var tabBarCircleSize = CGSize(width: 65, height: 65)
-    public static var tabBarSizeImage: CGFloat = 25
-    public static var tabBarShadowColor = UIColor.lightGray.cgColor
-    public static var tabBarSizeSelectedImage: CGFloat = 20
-    public static var tabBarAnimationDurationTime: Double = 0.4
-    
-}
-
-
 // use this protocol to detect when a tab bar item is pressed
 @available(iOS 10.0, *)
 protocol SOTabBarDelegate: AnyObject {
